refactor: useVerifiedLocations hook
