+++ src/services/responses/index.ts
@@ -1,0 +1,44 @@
+import { transformFoodResponse, FoodData } from "./food";
+import { transformBabalaResponse, BabalaData } from "./babala";
+import { transformAhbapResponse, AhbapData } from "./ahbap";
+import { transformHospitalResponse, HospitalData } from "./hospital";
+import { transformTeleteyitResponse, TeleteyitData } from "./teleteyit";
+import { transformSatelliteResponse, SatelliteData } from "./satellite";
+import { transformSahraResponse, SahraData } from "./sahra";
+import { transformPharmacyResponse, PharmacyData } from "./pharmacy";
+import { transformSafePlaceResponse, SafePlaceData } from "./safe-place";
+import { transformTwitterResponse, TwitterData } from "./twitter";
+import { transformTeyitEnkazResponse, TeyitEnkazData } from "./teyit-enkaz";
+import { transformTeyitYardimResponse, TeyitYardimData } from "./teyit-yardim";
+import { transformDepremIOResponse, DepremIOData } from "./depremio";
+
+export {
+  transformFoodResponse,
+  transformBabalaResponse,
+  transformAhbapResponse,
+  transformHospitalResponse,
+  transformTeleteyitResponse,
+  transformSatelliteResponse,
+  transformSahraResponse,
+  transformPharmacyResponse,
+  transformSafePlaceResponse,
+  transformTwitterResponse,
+  transformTeyitEnkazResponse,
+  transformTeyitYardimResponse,
+  transformDepremIOResponse,
+};
+export type {
+  FoodData,
+  BabalaData,
+  AhbapData,
+  HospitalData,
+  TeleteyitData,
+  SatelliteData,
+  SahraData,
+  PharmacyData,
+  SafePlaceData,
+  TwitterData,
+  TeyitEnkazData,
+  TeyitYardimData,
+  DepremIOData,
+};
